requirements.txt
torch=2.0.1
torchmetric=1.0.1
torch-geometric=
torch-scatter=2.1.1.
torch-cluster=1.6.1+pt20cu117
pytorch-lightning=2.0.3
